Update Threading Semantics
This commit makes a few changes related to the threading setup in `ConsumerApi` and `ProducerApi`.

It addresses unclean shutdown of the `java.util.concurrent.ExecutorService` as well as fixing a potential deadlock scenario.

Prior to this commit an `ExecutorService` with a single thread was created in `ConsumerApi.BlockingContext`. For blocking IO operations an unbounded thread pool is recommended (https://typelevel.org/cats-effect/concurrency/basics.html#choosing-thread-pool). Preventing unbounded resource usage is the responsibility for calling code running on a bounded thread pool.

This commit also updates the `ThreadFactory` so that if multiple pools happen to be created, they will be given globally unique names.

Two new methods are added to the companion object of `ConsumerApi` to allow for the caller to provider their own `Blocker` if so desired.

Also, the shutdown of the thread pool prior to this commit only called `(es: ExecutorService).shutdown()`. This is not sufficient to shutdown an `ExecutorService`. The full process is actually quite involved (https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/util/concurrent/ExecutorService.html). Thankfully `cats-effect` already provides this logic for us out of the box with `Blocker.fromExecutorService`.

Both the single threaded nature of the blocker _and_ the shutdown semantics were issues. The impetus for this commit is that we were experiencing a deadlock around shutdown/restart of a `fs2.Stream` with Kafka4s related code. I strongly believe that one or both of these is the source of that issue, however even if it is not, these items should still be addressed.

This is a binary incompatible change.