fix: remove library dependency on `github4s`
Having this as a library dependency tangles the versions of its dependencies
with the versions of `kafka4s`'s dependencies. The immediate problem is that
this is making the Cats Effect 3 upgrade intractable for no reason, but more
broadly speaking, we don't even care at the end of the day whether `github4s` is
implemented in Scala. We just need it in our build. Our dependencies should be
utterly independent of it.