feat(consumer): drop `wakeup` from `ConsumerApi`
Adapted from a conversation I had with @zcox: `wakeup` can be called safely by
any thread, and it will interrupt any operation that single thread is currently
performing. We used to rely on wakeup, but stopped using it once we had
`ShiftingConsumerImpl`. `wakeup` has nothing to do with reading records from
Kafka or committing offsets; It’s to help users deal with concurrency when using
the non-threadsafe Java client. Its presence on the interface is an
inconsistency in the level of abstraction; for us to expose it is to leak a
detail of the underlying Java client that is not relevant to the Scala code.
Moreover, it becomes possible to write all of `ShiftingConsumerImpl` as a
`.mapK` when we remove this odd man out (subsequent PR).