fix: referential opaqueness in Avro4s, as illustrated by Vulcan
Note that when a Vulcan Codec tries to generate the schema AST, its type
indicates that it can fail. This is true to the semantics; Avro4s does not
capture these failures at the type level, but being based on the same Avro Java
code underneath, the same things are true of it. We have seen in practice that
`SF.schema(DefaultFieldMapper)` can throw; and not only that, but it throws
sometimes, and not other times, even with the same
code (nondeterminism?). (Perhaps this should be wrapped with a `Sync` rather
than just an `ApplicativeThrow`?)